# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Digital ABD API
  version: 0.4.0
  description: |
    As-Built Documentation platform for Optical Fiber Cable (OFC) networks.

    **Authentication:** Bearer JWT (except `/health`, `/api/v1/auth/login`, `/api/v1/auth/oidc/*`).

    **Idempotency:** Include `Idempotency-Key` header on POST/PUT write operations.

    **Errors:** RFC 7807 Problem Details (`application/problem+json`).

    Live interactive docs available at `/docs` when the API server is running.
  contact:
    name: Digital ABD Program Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.digiabd.example.com
    description: Production (placeholder)
...
paths:
...
  /api/v1/routes/{routeId}/segments:
    get:
      tags: [Segments]
      summary: List segments for route
      operationId: listSegments
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/routeId'
      responses:
        '200':
          description: Segment list
    post:
      tags: [Segments]
      summary: Create segment
      operationId: createSegment
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/routeId'
        - name: Idempotency-Key
          in: header
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSegmentRequest'
      responses:
        '201':
          description: Segment created
  /api/v1/segments/{segmentId}:
    get:
      tags: [Segments]
      summary: Get segment by ID
      operationId: getSegment
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/segmentId'
      responses:
        '200':
          description: Segment
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/segments/{segmentId}/trench:
    put:
      tags: [Field Capture]
      summary: Upsert trench record
      operationId: upsertTrench
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/segmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrenchRecord'
      responses:
        '200':
          description: Trench saved
  /api/v1/segments/{segmentId}/hdd-crossing:
    put:
      tags: [Field Capture]
      summary: Upsert HDD crossing details
      operationId: upsertHddCrossing
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/segmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entry_latitude, entry_longitude, exit_latitude, exit_longitude, bore_length_m]
              properties:
                entry_latitude: { type: number }
                entry_longitude: { type: number }
                exit_latitude: { type: number }
                exit_longitude: { type: number }
                bore_length_m: { type: number }
                depth_m: { type: number }
                pipe_spec: { type: string }
      responses:
        '200':
          description: HDD crossing saved
...
  /api/v1/closures/{closureId}/otdr:
    post:
      tags: [Field Capture]
      summary: Upload OTDR test file for closure
      operationId: uploadOtdrTest
      security: [{ bearerAuth: [] }]
      parameters:
        - name: closureId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: OTDR test uploaded
  /api/v1/segments/{segmentId}/photos:
    post:
      tags: [Photos]
      summary: Upload geo-tagged photo
      operationId: uploadPhoto
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/segmentId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                phase: { type: string, enum: [before, during, after] }
                latitude: { type: number }
                longitude: { type: number }
      responses:
        '201':
          description: Photo uploaded
        '413':
          description: File exceeds 25 MB limit
...
  /api/v1/deviations:
    post:
      tags: [Deviations]
      summary: Create deviation record
      operationId: createDeviation
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeviationRequest'
      responses:
        '201':
          description: Deviation created
  /api/v1/deviations/{deviationId}/approve:
    post:
      tags: [Deviations]
      summary: Approve, reject, or return deviation
      operationId: approveDeviation
      security: [{ bearerAuth: [] }]
      parameters:
        - name: deviationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecision'
      responses:
        '200':
          description: Decision recorded
        '409':
          description: Invalid approval stage transition
  /api/v1/noc/lookup:
    get:
      tags: [NOC]
      summary: Fault localization asset lookup
      operationId: nocLookup
      security: [{ bearerAuth: [] }]
      parameters:
        - name: segment_id
          in: query
          schema: { type: string, format: uuid }
        - name: chainage
          in: query
          schema: { type: number }
        - name: latitude
          in: query
          schema: { type: number }
        - name: longitude
          in: query
          schema: { type: number }
      responses:
        '200':
          description: Asset context for fault ticket
...
  /api/v1/versions/{entityType}/{entityId}:
    get:
      tags: [Audit]
      summary: List version history for entity
      operationId: listRecordVersions
      security: [{ bearerAuth: [] }]
      parameters:
        - name: entityType
          in: path
          required: true
          schema: { type: string }
        - name: entityId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Version history list
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from /api/v1/auth/login or OIDC flow
  parameters:
...
    routeId:
      name: routeId
      in: path
      required: true
      schema: { type: string, format: uuid }
    segmentId:
      name: segmentId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
    CreateSegmentRequest:
      type: object
      required: [chainage_start, chainage_end]
      properties:
        chainage_start: { type: number }
        chainage_end: { type: number }
        surface_type: { type: string, enum: [road, shoulder, rural, urban, highway, railway] }
    TrenchRecord:
      type: object
      required: [depth_m]
      properties:
        depth_m: { type: number, minimum: 0, example: 1.65 }
        width_m: { type: number }
        bedding_type: { type: string }
        reinstatement_status: { type: string, enum: [pending, in_progress, completed] }
...
    CreateDeviationRequest:
      type: object
      required: [segment_id, category, description]
      properties:
        segment_id: { type: string, format: uuid }
        category: { type: string }
        description: { type: string }
        justification: { type: string }
        severity: { type: string, enum: [low, medium, high, critical] }
    ApprovalDecision:
      type: object
      required: [decision]
      properties:
        decision: { type: string, enum: [approved, rejected, returned] }
        comments: { type: string }
...